requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic>=2.8.0
email-validator==2.1.0.post1
python-dotenv==1.0.0
requests==2.31.0
PyPDF2==3.0.1
langsmith>=0.1.77
langchain-google-genai>=2.1.0
langchain>=0.3.0

